// src/resources/verification_codes/enums/code-messages.ts
export enum CodeMessages {
  INVALID_CODE = 'Invalid verification code.',
  SENT = 'A verification code sent',
  ALREADY_SENT = 'A verification code have already sent.',
  CODE_SENT_TO_EMAIL = 'A verification code sent to your email address.',
  CODE_SENT_TO_MOBILE_PHONE = 'A verification code sent to your mobile phone.',
}